--- packages/dapp/src/components/DualWidget/index.ts
@@ -1,1 +1,0 @@
-export { DualWidget } from './DualWidget';
