+++ packages/dapp/src/components/StatsModal/StatsModal.style.ts
@@ -1,10 +1,56 @@
-import { AppBar, AppBarProps, Box } from '@mui/material';
+import {
+  AppBar,
+  AppBarProps,
+  Box,
+  Breakpoint,
+  Grid,
+  GridProps,
+} from '@mui/material';
 
 import { alpha, styled } from '@mui/material/styles';
 
 export interface ModalHeaderAppBarProps extends Omit<AppBarProps, 'component'> {
   component?: string;
 }
+
+export const ModalContainer = styled(AppBar)<ModalHeaderAppBarProps>(
+  ({ theme }) => ({
+    position: 'absolute',
+    top: '50%',
+    left: '50%',
+    transform: 'translate(-50%, -50%)',
+    margin: 'auto',
+    paddingBottom: theme.spacing(6),
+    borderRadius: '12px',
+    boxShadow:
+      theme.palette.mode === 'dark'
+        ? '0px 2px 4px rgba(0, 0, 0, 0.08), 0px 8px 16px rgba(0, 0, 0, 0.16)'
+        : '0px 2px 4px rgba(0, 0, 0, 0.08), 0px 8px 16px rgba(0, 0, 0, 0.08)',
+    width: `calc( 100% - ${theme.spacing(6)})`,
+    maxWidth: '640px',
+    maxHeight: '85%',
+    overflowY: 'auto',
+    background:
+      theme.palette.mode === 'dark'
+        ? theme.palette.surface2.main
+        : theme.palette.surface1.main,
+
+    [theme.breakpoints.up('md' as Breakpoint)]: {
+      margin: 0,
+      width: '640px',
+    },
+  }),
+);
+
+export const ModalContent = styled(Grid)<GridProps>(({ theme }) => ({
+  alignItems: 'center',
+  gap: '28px',
+  padding: theme.spacing(0, 6),
+  justifyContent: 'space-between',
+  [theme.breakpoints.up('md' as Breakpoint)]: {
+    justifyContent: 'inherit',
+  },
+}));
 
 export const ModalHeaderAppBar = styled(AppBar)<ModalHeaderAppBarProps>(
   ({ theme }) => ({
