+++ packages/dapp/src/components/StatsModal/StatsModal.tsx
@@ -1,19 +1,19 @@
 import CloseIcon from '@mui/icons-material/Close';
 import {
   Avatar,
-  Box,
-  Breakpoint,
   Grid,
   IconButton,
   Modal as MUIModal,
   Typography,
-  useMediaQuery,
   useTheme,
 } from '@mui/material';
-import { ModalHeaderAppBar } from './StatsModal.style';
+import {
+  ModalContainer,
+  ModalContent,
+  ModalHeaderAppBar,
+} from './StatsModal.style';
 export const StatsModal = ({ title, open, setOpen, data }) => {
   const theme = useTheme();
-  const isDesktop = useMediaQuery(theme.breakpoints.up('md' as Breakpoint));
   return (
     <MUIModal
       open={open}
@@ -27,29 +27,7 @@ export const StatsModal = ({ title, open, setOpen, data }) => {
       aria-labelledby="modal-modal-title"
       aria-describedby="modal-modal-description"
     >
-      <Box
-        sx={{
-          position: 'absolute',
-          top: '50%',
-          left: '50%',
-          transform: 'translate(-50%, -50%)',
-          margin: isDesktop ? 0 : 'auto',
-          paddingBottom: theme.spacing(6),
-          borderRadius: '12px',
-          boxShadow:
-            theme.palette.mode === 'dark'
-              ? '0px 2px 4px rgba(0, 0, 0, 0.08), 0px 8px 16px rgba(0, 0, 0, 0.16)'
-              : '0px 2px 4px rgba(0, 0, 0, 0.08), 0px 8px 16px rgba(0, 0, 0, 0.08)',
-          width: isDesktop ? '640px' : `calc( 100% - ${theme.spacing(6)})`,
-          maxWidth: '640px',
-          maxHeight: '85%',
-          overflowY: 'auto',
-          background:
-            theme.palette.mode === 'dark'
-              ? theme.palette.surface2.main
-              : theme.palette.surface1.main,
-        }}
-      >
+      <ModalContainer>
         <ModalHeaderAppBar>
           <Typography
             id="modal-modal-title"
@@ -79,14 +57,8 @@ export const StatsModal = ({ title, open, setOpen, data }) => {
             <CloseIcon />
           </IconButton>
         </ModalHeaderAppBar>
-        <Grid
-          container
-          alignItems="center"
-          gap={'28px'}
-          p={theme.spacing(0, 6)}
-          justifyContent={isDesktop ? 'inherit' : 'space-between'}
-        >
-          {data.map((el, index) => {
+        <ModalContent container>
+          {data?.map((el, index) => {
             return (
               <Grid
                 item
@@ -118,8 +90,8 @@ export const StatsModal = ({ title, open, setOpen, data }) => {
               </Grid>
             );
           })}
-        </Grid>
-      </Box>
+        </ModalContent>
+      </ModalContainer>
     </MUIModal>
   );
 };
