+++ packages/dapp/src/components/Widgets/index.ts
@@ -1,0 +1,1 @@
+export { Widgets } from './Widgets';
