+++ packages/dapp/src/components/index.ts
@@ -1,5 +1,4 @@
 export * from './BackgroundGradient';
-export * from './DualWidget';
 export * from './Navbar';
 export * from './OnRamper';
 export * from './StatsCard';
@@ -8,3 +7,4 @@ export * from './SupportModal';
 export * from './ThemeSwitch';
 export * from './WelcomeWrapper';
 export * from './Widget';
+export * from './Widgets';
