+++ packages/dapp/src/hooks/index.ts
@@ -1,1 +1,2 @@
+export * from './useFetchDexsAndBridges';
 export * from './userTracking';
