+++ packages/shared/src/fonts/gt-america.css
@@ -1,12 +1,11 @@
-@font-face {
+/* @font-face {
   font-family: GT-America;
   font-style: bold;
   font-weight: 800;
   font-display: swap;
   src: url('GT-America-Black.woff2') format('woff2'),
     url('GT-America-Black.woff') format('woff');
-}
-
+} */
 
 @font-face {
   font-family: GT-America;
@@ -16,7 +15,7 @@
   src: url('GT-America-Bold.woff2') format('woff2'),
     url('GT-America-Bold.woff') format('woff');
 }
-
+/* 
 @font-face {
   font-family: GT-America;
   font-style: normal;
@@ -24,7 +23,7 @@
   font-display: swap;
   src: url('GT-America-Medium.woff2') format('woff2'),
     url('GT-America-Medium.woff') format('woff');
-}
+} */
 
 @font-face {
   font-family: GT-America;
@@ -34,7 +33,7 @@
   src: url('GT-America-Regular.woff2') format('woff2'),
     url('GT-America-Regular.woff') format('woff');
 }
-
+/* 
 @font-face {
   font-family: GT-America;
   font-style: normal;
@@ -60,4 +59,4 @@
   font-display: swap;
   src: url('GT-America-Thin.woff2') format('woff2'),
     url('GT-America-Thin.woff') format('woff');
-}
+} */
